projects/fogg/calendar/pyproject.toml
[project]
name = "calendar"
version = "0.1.0"
description = "FOGG Calendar Automation with Google Groups Integration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic[vertex]>=0.58.2",
    "google-api-python-client>=2.176.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "mcp>=1.12.0",
    "pydantic>=2.11.7",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "python-dateutil>=2.9.0.post0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "ruff>=0.12.4",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "C4", "SIM", "RUF"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
